Implement foldLeft function for Event
Event now has a foldLeft function that takes an initial value and a
function to apply to each following occurrence (taking the last and the
current occurrence values as input).

This function should allow for counters/accumulators to be created from
Events and also should allow for rudimentary integration of Behaviours
using sampling.