Cache pull Events
Map/merge Events now cache their last evaluated value so the map and
merge are not recomputed.

To simplify this a Cache class was implemented. This basically allows
for a pure monotonically evaluated function to be cached easily. The
Cache class has also been utilised in the Behaviour implementation to
clean it up.